agents/coach-companion/requirements.txt
# Strands Agents SDK
strands-agents==0.1.0
strands-agents-tools==0.1.0

# Web framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic>=2.11.0

# AWS Bedrock (included in strands-agents base)
boto3==1.34.34

# Utilities
python-dotenv==1.0.0
